The `Fiber.current ()` operation must never propagate cancelation
The problem is that in order to call `Fiber.forbid fiber (fun () -> ...)` to
forbid cancelation, it is necessary to call `Fiber.current ()` to obtain the
`fiber`.  Hence `Fiber.current ()` must not propagate cancelation.  In some
cases one could call `Fiber.current ()` well before performing the work that
needs to be protected by `Fiber.forbid ...`, but this is not always practical.