Allow resume from signal handlers
While
[`pthread_cond_broadcast`](https://pubs.opengroup.org/onlinepubs/007904975/functions/pthread_cond_broadcast.html)
is not safe to call from a signal handler, the situation in OCaml is somewhat
different as [signal handlers in OCaml are called during poll
points](https://ocaml.github.io/ocamlunix/signals.html#sec86) on some thread.

The change in here makes an assumption that there is no poll point in
`Condition.wait` before the thread has been suspended and can be resumed by a
`Condition.broadcast`.  If this is the case, then it should actually be
possible, under the right circumstances, to use `Condition.broadcast` from a
signal handler to release a waiting thread.