When `threads.posix` exists we really need `thread-local-storage`
`thread-local-storage` is needed for correct operation when `threads.posix`
exists.  TLS is used to store the fiber identity with `picos.threaded` and that
breaks without proper TLS.

I've now run into this a couple of times that I missed `thread-local-storage`
from the switch and that unfortunately allowed an invalid build where
`threads.posix` (and hence systhreads) existed without `thread-local-storage`.