Change `Picos_structured.Run.any []` to not block
There are a number of reasons for making this change.

While it sort of makes logical sense to block on an empty list, that can also be
surprising and undesirable.  This has also come up as an opinion in private
communications.

For example, suppose you want to use `Run.any` to start actions to try to search
for a solution concurrently and the list of actions to use in the search is
dynamically created.  A simple program such as

```ocaml
let solution = Atomic.make None in
...
Run.any search_actions;
match Atomic.get solution with
| None -> ...
| Some solution -> ...
```

might unexpectedly block in case the list of actions would be empty.

The above also illustrates that it is typically not difficult to implement the
blocking semantics on top of the non-blocking semantics (i.e. just block in the
`None` case).

It is still allowed to cancel individual fibers started with `Run.any`.
Consider

```ocaml
Run.any []
```

and

```ocaml
Run.any [ fun () -> raise Control.Terminate ]
```

Is there a difference between the two?  Should both block and, if so, does that
open more possibilities to write unexpectedly blocking code?

OTOH, disallowing canceling fibers used with `Run.any` (and `Run.all`) by
raising `Control.Terminate` would seem to introduce an inconsistency with
respect to other constructs in `Picos_structured`.

I'm happy to revisit this later in case more motivating examples arise.