Try to reproduce SIGSEGV
I noticed the server and client test crashed with a SIGSEGV or SIGBUS (depending
on run).  I created a stripped down version of the program

  https://github.com/polytypic/signal-issue

that removes all "unsafe" optimizations and it still crashes on macOS with
ARM (tested on M1 and M3).

I recall seeing the server and client test also crash on non macOS machines, but
it seems to be more difficult to reproduce.  This is an ettempt to get it to
reproduce on OCaml-CI.