Fix to fully transmit body
I noticed that when using the bytecode interpreter the request/response body was
not always properly transmitted --- a test will appear in a separate PR.  This
simplifies the client to not try to write things concurrently and changes both
sides to shutdown the connection partially after being done with writing or
reading.  This seems to have eliminated the issue.