Split Picos into multiple packages and remove `Exn_bt`
This splits the picos package into several smaller packages, roughly:

- `picos` -> (`backoff`, `thread-local-storage`)
- `picos_aux` -> (`backoff`, `multicore-magic`)
- `picos_std` -> `picos`(, `backoff`, `multicore-magic`)
- `picos_stdio` -> `picos_std`, `picos_aux`, and `unix`(, `mtime`, `psq`)
- `picos_stdio_cohttp` -> `picos_stdio`(, `cohttp`, `fmt`, `uri`)
- `picos_mux` -> `picos_std`, `picos_aux`, and optionally on `picos_stdio`
- `picos_lwt` -> `picos_std`, `picos_aux`(, `lwt`)

- `picos_meta` -> all of the above for integration tests and to simplify
  dependencies of other packages

The above changes should make it easier for people to depend only on the
packages they need.  A single package with dependencies to `lwt` and `cohttp` is
just too heavy.

This also removes the `Exn_bt` module and type entirely.  It is impossible to
have such a module in OCaml.

Due to the split to multiple packages the structure of the documentation has
also been updated.  Currently odoc doesn't support references between packages,
which, unfortunately, downgrades the quality of the documentation.